Changed the parsing functions.
Modified the way in which the words, definitions and stopwords are
unified in one object. A word is now a key in a map called "wordDefs"
which points to an array of all the definitions that belong to this
word. Stopwords are now forged into a single regular expression string.